+++ app/src/main/java/woowacourse/paint/presentation/ui/canvas/CanvasView.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.presentation.ui.canvas
 
 import android.content.Context
 import android.graphics.Canvas
