+++ app/src/main/java/woowacourse/paint/presentation/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.presentation.ui.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
