+++ app/src/main/res/layout/activity_main.xml
@@ -17,7 +17,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <woowacourse.paint.CanvasView
+        <woowacourse.paint.presentation.ui.canvas.CanvasView
             android:id="@+id/cvCanvas"
             android:layout_width="0dp"
             android:layout_height="0dp"
